Use filter notation
It is not necessary to use the code in the old form. Instead, the filter integrates JSXGraph. In addition, the div does not need to be created separately, only the usual filter notation is used. The JavaScript code does not need to be packaged in a jQuery function. The filter ensures execution at the end of the loading process.