In this new version, our filter makes another big leap forward. From now on it is also possible to use several boards in the fomulas extension `JSXQuestion`. For this purpose, the operation of the class has been significantly expanded. Of course, the old way of using it still works.
In order to use several boards in a form question from now on, pass the array `BOARDIDS` at the initialization. Also make sure that you *not* use `JXG.JSXGraph.initBoard`, but the `initBoards` function of the JSXQuestion class.

Further innovations for several boards:
- Use the attribute `BOARDIDS` instead of `BOARDID`
- Use the attribute `boards` instead of `board`
- Use the `initBoards` method instead of `initBoard`

Generally, at the end of the function for the JSXGraph code, there is no longer any need to update the board.